portfolio/01-defect-detection/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
fastapi>=0.100.0
uvicorn>=0.23.0
pillow>=10.0.0
albumentations>=1.3.0
opencv-python>=4.8.0
onnxruntime>=1.15.0
python-multipart>=0.0.6
matplotlib>=3.7.0
numpy>=1.24.0
